PRODX-20593 WIP: Move Connection to separate element
